move work on classes back to afskmodem.py
