+++ afskmodem.py
@@ -8,13 +8,15 @@ import wave
 import struct
 import pyaudio
 from datetime import datetime
-import os
 from time import sleep
 
-##### Constants #####
+# ----- Constants -----
 
-##### USER PARAMETERS: Although these should work fine as-is, 
-##### tuning them will not affect functionality.
+# USER PARAMETERS: Although these should work fine as-is, 
+# tuning them will not affect functionality.
+
+# Log level (0: Info (Recommended), 1: Warn, 2: Error, 3: None)
+LOG_LEVEL = 0
 
 # Training sequence time in seconds (0.5-1.0, Default 0.6)
 TRAINING_SEQUENCE_TIME = 0.6
@@ -28,8 +30,7 @@ AMPLITUDE_END_THRESHOLD = 14000
 # Amplifier function deadzone (0-32768, Default 128 [-48.2 dBfs])
 AMPLIFIER_DEADZONE = 128
 
-
-##### PROGRAM CONSTANTS: DO NOT CHANGE THESE!
+# PROGRAM CONSTANTS: Do not change these!
 
 # Frames per buffer for audio input (Default 2048 [0.043s])
 INPUT_FRAMES_PER_BLOCK = 2048
@@ -53,53 +54,243 @@ WAVEFORMS_DIR = "waveforms/"
 """
     Logging functionality
 """
-def get_date_and_time(): # Long date and time for logging
-        return datetime.now().strftime('%Y-%m-%d %H:%M:%S')
-
-# Logging level (0: INFO, 1: WARN (recommended), 2: ERROR, 3: NONE)
-LOG_LEVEL = 0
-#
-# Should the log output to the console?
-LOG_TO_CONSOLE = True
-#
-# Should the log output to a log file?
-LOG_TO_FILE = False
-#
-# Where to generate logfile if need be
-LOG_PATH = "afskmodem.log"
-#
-# How the log identifies which module is logging.
-LOG_PREFIX = "(AFSKmodem)"
-
-# Initialize log file if needed
-if(LOG_TO_FILE):
-    try:
-        os.remove(LOG_PATH)
-    except:
-        pass
-    with open(LOG_PATH, "w") as f:
-        f.write(get_date_and_time() + 
-                " [  OK  ] " + LOG_PREFIX + " Logging initialized.\n")
-
-# Prints a message to enabled log outputs with the given level.
-def log(level: int, data: str):
+# Logs an event
+def log(level: int, message: str):
     if(level >= LOG_LEVEL):
-        output = get_date_and_time()
+        output = datetime.now().strftime('%Y-%m-%d %H:%M:%S')
         if(level == 0):
             output += " [  OK  ] "
         elif(level == 1):
             output += " [ WARN ] "
         else:
             output += " [ERROR!] "
-        output += LOG_PREFIX + " "
-        output += data
-        if(LOG_TO_FILE):
-            with open(LOG_PATH, "a") as f:
-                f.write(output + "\n")
-        if(LOG_TO_CONSOLE):
-            print(output)
+        output += "(AFSKmodem): "
+        output += message
+        print(output)
 
+"""
+    DigitalModes defines the speeds at which AFSKmodem can operate, 
+    and the tones that they use.
+"""
+class DigitalModes:
+    # Audio Frequency-Shift Keying (300 baud)
+    def afsk300() -> str: 
+        return "afsk300"
 
+    # Audio Frequency-Shift Keying (600 baud)
+    def afsk600() -> str: 
+        return "afsk600"
+    
+    # Audio Frequency-Shift Keying (1200 baud)
+    def afsk1200() -> str: 
+        return "afsk1200"
+
+    # Audio Frequency-Shift Keying (2400 baud)
+    def afsk2400() -> str: 
+        return "afsk2400"
+
+    # Audio Frequency-Shift Keying (6000 baud)
+    def afsk6000() -> str: 
+        return "afsk6000"
+
+    # Default (AFSK1200)
+    def default() -> str: 
+        return "afsk1200"
+    
+    # Returns the unit time in samples of the given mode
+    def getUnitTime(digital_mode: str) -> int:
+        if(digital_mode == "afsk300"):
+            return int(SAMPLE_RATE / 300)
+        elif(digital_mode == "afsk600"):
+            return int(SAMPLE_RATE / 600)
+        elif(digital_mode == "afsk1200"):
+            return int(SAMPLE_RATE / 1200)
+        elif(digital_mode == "afsk2400"):
+            return int(SAMPLE_RATE / 2400)
+        elif(digital_mode == "afsk6000"):
+            return int(SAMPLE_RATE / 6000)
+        else: # default
+            return int(SAMPLE_RATE / 1200)
+
+    # Returns the number of training sequence oscillations 
+    # for a specified time
+    def getTrainingCycles(sequence_time: int, digital_mode: str) -> int:
+        if(digital_mode == "afsk300"):
+            return int(300 * sequence_time / 2)
+        elif(digital_mode == "afsk600"):
+            return int(600 * sequence_time / 2)
+        elif(digital_mode == "afsk1200"):
+            return int(1200 * sequence_time / 2)
+        elif(digital_mode == "afsk2400"):
+            return int(2400 * sequence_time / 2)
+        elif(digital_mode == "afsk6000"):
+            return int(6000 * sequence_time / 2)
+        else: # default
+            return int(1200 * sequence_time / 2)
+    
+    # Get the frequency of the space tone for a given mode
+    def getSpaceTone(digital_mode: str) -> int:
+        if(digital_mode == "afsk300"):
+            return 300
+        elif(digital_mode == "afsk600"):
+            return 600
+        elif(digital_mode == "afsk1200"):
+            return 1200
+        elif(digital_mode == "afsk2400"):
+            return 2400
+        elif(digital_mode == "afsk6000"):
+            return 6000
+        else: # default
+            return 1200
+
+    # Get the frequency of the mark tone for a given mode
+    def getMarkTone(digital_mode: str) -> int:
+        if(digital_mode == "afsk300"):
+            return 600
+        elif(digital_mode == "afsk600"):
+            return 1200
+        elif(digital_mode == "afsk1200"):
+            return 2400
+        elif(digital_mode == "afsk2400"):
+            return 4800
+        elif(digital_mode == "afsk6000"):
+            return 12000
+        else: # default
+            return 2400
+"""
+    Waveforms provides functions to load waveforms from files and return them 
+    as lists of amplitudes, stored as either bytes or ints.
+"""
+class Waveforms:
+    def __init__(self, digital_mode):
+        self.digital_mode = digital_mode
+
+    # Load waveform data from a file 
+    # and convert it to a list of amplitudes as ints
+    def __loadWavFile(self, filename: str) -> list[int]:
+        with wave.open(filename, "r") as f:
+            nFrames = f.getnframes()
+            expFrames = []
+            for i in range(0, nFrames):
+                sFrame = f.readframes(1)
+                expFrames.append(struct.unpack("<h", sFrame)[0])
+            return expFrames
+
+    # Load waveform data from a file 
+    # and convert it to a list of amplitudes as bytes
+    def __loadRawWavFile(self, filename: str) -> bytes:
+        with wave.open(filename, "r") as f:
+            nFrames = f.getnframes()
+            return f.readframes(nFrames)
+    
+    # Silence (20ms) to pad output with for TX
+    def getTxSilence(self) -> bytes: 
+        return self.__loadRawWavFile(WAVEFORMS_DIR + "_.wav")
+    
+    # Space tone as bytes for transmitting
+    def getTxSpace(self) -> bytes: 
+        return self.__loadRawWavFile(WAVEFORMS_DIR + 
+                                     self.digital_mode + "/0.wav")
+    
+    # Mark tone as bytes for transmitting
+    def GetTxMark(self) -> bytes: 
+        return self.__loadRawWavFile(WAVEFORMS_DIR + 
+                                     self.digital_mode + "/1.wav")
+    
+    # Space tone as int array for receiving
+    def getRxSpace(self) -> list[int]: 
+        return self.__loadWavFile(WAVEFORMS_DIR + 
+                                  self.digital_mode + "/0.wav")
+    
+    # Mark tone as int array for receiving
+    def getRxMark(self) -> list[int]: 
+        return self.__loadWavFile(WAVEFORMS_DIR + 
+                                  self.digital_mode + "/1.wav")
+    
+    # Single training sequence cycle for RX clock recovery
+    def getRxTraining(self) -> list[int]: 
+        return self.getRxMark() + self.getRxSpace()
+
+"""
+    ECC provides functionality for encoding, decoding, and correcting
+    bit errors.
+"""
+class ECC:
+    # Pad the positions of parity bits in a byte with 0s
+    def __padParityBits(bits: str) -> str:
+        j = 0
+        k = 1
+        pad_bits = ""
+        for i in range(1, len(bits) + 5):
+            if(i == 2**j):
+                pad_bits += '0'
+                j += 1
+            else:
+                pad_bits += bits[-1 * k]
+                k += 1
+        return pad_bits[::-1]
+
+    # Trim the parity bits off an encoded byte
+    def __trimParityBits(bits: str) -> str:
+        bits = bits[::-1]
+        j = 0
+        trimmed_bits = ""
+        for i in range(len(bits)):
+            if(i + 1 != 2 ** j):
+                trimmed_bits += bits[i]
+            else:
+                j += 1
+        return trimmed_bits[::-1]
+
+    # Find an error (if it exists) and return the index
+    def __getErrorIndex(bits: str) -> int:
+        n = len(bits)
+        p = 0
+        for i in range(4):
+            k = 0
+            for j in range(1, n + 1):
+                if(j & (2**i) == (2**i)):
+                    k = k ^ int(bits[-1 * j])
+            p += k*(10**i)
+        return n - int(str(p), 2)
+ 
+    # Encodes a byte of data (8 bits -> 12 bits)
+    def __encodeByte(bits: str) -> str:
+        p_bits = ECC.__padParityBits(bits)
+        n = len(p_bits)
+        for i in range(4):
+            p = 0
+            for j in range(1, n + 1):
+                if(j & (2**i) == (2**i)):
+                    p = p ^ int(p_bits[-1 * j])
+            p_bits = p_bits[:n-(2**i)] + str(p) + p_bits[n-(2**i)+1:]
+        return p_bits
+
+    # Decodes a byte of data (12 bits -> 8 bits)
+    def __decodeByte(bits: str) -> (str, bool):
+        err_index = ECC.__getErrorIndex(bits)
+        if(err_index == len(bits)):
+            return (ECC.__trimParityBits(bits), False)
+        else:
+            bits[err_index] = "0" if bits[err_index] == "1" else "1"
+        return (ECC.__trimParityBits(bits), True)
+
+    # Corrects errors and decodes data. Returns decoded bits & number of errors corrected.
+    def decode(bits: str) -> (str, int):
+        error_count = 0
+        dec_bytes = ""
+        for i in range(0, len(bits) - 11, 12):
+            (dec_byte, err_found) = ECC.__decodeByte(bits[i : i + 1])
+            dec_bytes += dec_byte
+            error_count += 1 if err_found else 0
+        return (dec_bytes, error_count)
+    
+    # Encodes data by inserting parity bits
+    def encode(bits: str) -> str:
+        enc_bits = ""
+        for i in range(0, len(bits) - 7, 8):
+            enc_bits += ECC.__encodeByte(bits[i:i+8])
+        return enc_bits
 
 """
     DigitalReceiver is a user-defined receiver class that provides functionality 
@@ -118,11 +309,10 @@ class DigitalReceiver:
         self.unit_time = DigitalModes.getUnitTime(self.digital_mode)
         self.space_tone = DigitalModes.getSpaceTone(self.digital_mode)
         self.mark_tone = DigitalModes.getMarkTone(self.digital_mode)
-        ideal_waves = IdealWaves(self.digital_mode)
-        self.rx_space = ideal_waves.getRxSpace()
-        self.rx_mark = ideal_waves.getRxMark()
-        self.rx_training = ideal_waves.getRxTraining()
-        self.ecc = Hamming()
+        waveforms = Waveforms(self.digital_mode)
+        self.rx_space = waveforms.getRxSpace()
+        self.rx_mark = waveforms.getRxMark()
+        self.rx_training = waveforms.getRxTraining()
     
     # Amplifies a chunk of audio for decoding
     def __amplifyChunk(self, chunk: list[int]) -> list[int]:
@@ -282,21 +472,21 @@ class DigitalReceiver:
         return bytes(int_data)
     
     # Receives data and returns it (or fails)
-    def rx(self, timeout = -1):
+    def rx(self, timeout: int = -1) -> (bytes, int):
         log(0, "Receiver - listening...")
-        recv_wav = self.__listen(timeout)
-        if(recv_wav == b""): # if timed out
+        recv_audio = self.__listen(timeout)
+        if(recv_audio == b""): # if timed out
             log(1, "Receiver - timed out.")
             return b"", 0
-        recv_bits = self.__decodeBits(recv_wav)
+        recv_bits = self.__decodeBits(recv_audio)
         if(recv_bits == ""): # if no good data
             log(1, "Receiver - bad data.")
             return b"", 0
         recv_bits = self.__trimTrainingSeq(recv_bits)
-        decoded_bin, error_count = self.__correctErrors(recv_bits)
-        bytes_data = self.__bitsToBytes(decoded_bin)
+        (dec_bits, error_count) = ECC.decode(recv_bits)
+        dec_bytes = self.__bitsToBytes(dec_bits)
         log(0, "Receiver - done.")
-        return bytes_data, error_count
+        return (dec_bytes, error_count)
 
 """
     DigitalTransmitter is a user-defined transmitter class that provides
@@ -310,11 +500,10 @@ class DigitalTransmitter:
         self.ts_cycles = DigitalModes.getTrainingCycles(
             training_sequence_time, self.digital_modulation_type)
         self.unit_time = DigitalModes.getUnitTime(self.digital_modulation_type)
-        ideal_waves = IdealWaves(digital_mode = self.digital_modulation_type)
-        self.tx_space = ideal_waves.getTxSpace()
-        self.tx_mark = ideal_waves.GetTxMark()
-        self.tx_silence = ideal_waves.getTxSilence()
-        self.ecc = Hamming()
+        waveforms = Waveforms(digital_mode = self.digital_modulation_type)
+        self.tx_space = waveforms.getTxSpace()
+        self.tx_mark = waveforms.GetTxMark()
+        self.tx_silence = waveforms.getTxSilence()
 
     # Convert bytes to bits
     def __bytesToBits(self, b_in: bytes) -> str:
@@ -345,9 +534,6 @@ class DigitalTransmitter:
         output += "0" * 3
         return output
 
-    # Generate ECC and insert into data.
-
-
     # Play a waveform
     def __playWaveform(self, data: bytes):
             # Open an output stream with PortAudio
@@ -369,13 +555,13 @@ class DigitalTransmitter:
     def tx(self, data: bytes): 
         log(0, "Transmitter - sending " + str(len(data)) + " bytes...")
         message_bits = self.__bytesToBits(data)
-        ecc_bits = self.__generateECC(message_bits)
+        ecc_bits = ECC.encode(message_bits)
         training_block = self.__getTrainingBlock()
         tx_bits = training_block + ecc_bits
-        out_frames = self.__encodeBits(tx_bits)
-        self.__playWaveform(out_frames)
+        tx_audio = self.__encodeBits(tx_bits)
+        self.__playWaveform(tx_audio)
         log(0, "Transmitter - done.")
     
     # Estimate transmission time in seconds
-    def estimateTxTime(self, length: int): 
+    def estimateTxTime(self, length: int) -> float: 
         return (self.ts_cycles * 2 + length * 12) / (SAMPLE_RATE / self.unit_time)
